move the Nucleus class to nucdata
