--- docs/source/pynucastro.nucleus.rst
@@ -1,18 +1,0 @@
-pynucastro.nucleus package
-==========================
-
-.. automodule:: pynucastro.nucleus
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-pynucastro.nucleus.nucleus module
----------------------------------
-
-.. automodule:: pynucastro.nucleus.nucleus
-   :members:
-   :undoc-members:
-   :show-inheritance:
