+++ docs/source/pynucastro.rst
@@ -14,6 +14,5 @@ Subpackages
 
    pynucastro.networks
    pynucastro.nucdata
-   pynucastro.nucleus
    pynucastro.rates
    pynucastro.screening
