+++ pynucastro/nucdata/__init__.py
@@ -11,3 +11,4 @@ from .elements import Element, UnidentifiedElement, PeriodicTable
 from .partition_function import PartitionFunction, PartitionFunctionTable, PartitionFunctionCollection
 from .spin_nuclide import SpinNuclide, SpinTable
 from .mass_nuclide import MassNuclide, MassTable
+from .nucleus import Nucleus, UnsupportedNucleus
