--- pynucastro/nucleus/__init__.py
@@ -1,8 +1,0 @@
-"""Properties of nuclei needed in construction reaction rates (in
-particular, the binding energies)
-
-"""
-
-__all__ = ["nucleus"]
-
-from .nucleus import Nucleus, UnsupportedNucleus
