+++ pynucastro/rates/library.py
@@ -2,7 +2,7 @@ import os
 import io
 import collections
 
-from pynucastro.nucleus import Nucleus, UnsupportedNucleus
+from pynucastro.nucdata import Nucleus, UnsupportedNucleus
 from pynucastro.rates import Rate, _find_rate_file
 from pynucastro.rates.rate import DerivedRate
